// src/roles/roles.enum.ts
/* eslint-disable prettier/prettier */
enum TravelRegRoles {
    Admin = 'Admin',
    SubAdmin = 'Sub-Admin',
    // AreaAdmin = 'Area-Admin',
    ParkAdmin = 'Park-Admin',
    // TransportCompany = 'Transport-Company',
    Driver = 'Driver',
    // Passenger = 'Passenger',
}
export default TravelRegRoles